Applied selectors to Detail screen.
