+++ pwa/src/pages/Detail/index.tsx
@@ -3,7 +3,6 @@ import { StyleSheet, Text, View, SafeAreaView } from 'react-native'
 import { TouchableOpacity } from 'react-native-gesture-handler'
 import { RouteProp, useRoute, useNavigation } from '@react-navigation/native'
 import { StackNavigationProp } from '@react-navigation/stack'
-import { AppLoading } from 'expo'
 import { Feather as Icon } from '@expo/vector-icons'
 
 import ReturnButton from '../../components/ReturnButton'
@@ -11,27 +10,21 @@ import MainStockInfo from './components/MainStockInfo'
 import TransactionList from './components/TransactionList'
 
 import API from '../../services/api'
-import { Transaction, Stock } from '../../services/api/types'
-import { YahooStock } from '../../services/yahooFinance/stockInfo'
+import { Transaction } from '../../services/api/types'
+
 import DataContext from '../../store/dataContext'
+import { getStockData } from '../../store/selectors'
 
 import { AppStackParamList } from '../../routes/AppStack'
-
 
 const Detail = () => {
     const [transactionList, setTransactionList] = useState<Transaction[]>([])
 
-    const {
-        state: {
-            stocksData,
-            isStocksDataReady,
-            yahooData,
-            isYahooDataReady
-        }
-    } = useContext(DataContext)
-
     const navigation = useNavigation<StackNavigationProp<AppStackParamList>>()
     const { params: { ticker } } = useRoute<RouteProp<AppStackParamList, 'Detail'>>()
+
+    const { state } = useContext(DataContext)
+    const stockData = getStockData(state, ticker)
 
     const navigateToNewTransactionWith = useCallback((type: 'IN' | 'OUT') => {
         navigation.navigate('NewTransaction', {
@@ -44,14 +37,6 @@ const Detail = () => {
         API.getTransactionsFor(ticker).then(setTransactionList)
     }, [])
 
-    if (!isStocksDataReady || !isYahooDataReady || !stocksData.has(ticker)) {
-        return <AppLoading />
-    }
-
-    const { currently_owned_shares, average_bought_price } = stocksData.get(ticker) as Stock
-
-    const { regularMarketPrice } = yahooData.get(ticker) as YahooStock
-
     return (
         <SafeAreaView style={styles.outerContainer}>
 
@@ -63,11 +48,14 @@ const Detail = () => {
                     <Text style={styles.title}>{ticker}</Text>
                 </View>
 
-                <MainStockInfo
-                    averageBoughtPrice={average_bought_price}
-                    currentlyOwnedShares={currently_owned_shares}
-                    regularMarketPrice={regularMarketPrice}
-                />
+                {
+                    stockData.regularMarketPrice &&
+                    <MainStockInfo
+                        averageBoughtPrice={stockData.average_bought_price}
+                        currentlyOwnedShares={stockData.currently_owned_shares}
+                        regularMarketPrice={stockData.regularMarketPrice}
+                    />
+                }
 
                 <View style={styles.buttonsContainer}>
                     <TouchableOpacity
